refactor: split tenant settings into Defaults page and standalone Features
Replaces the consolidated /tenants/{id}/settings page (with locale +
feature toggles inline) with two distinct entries under the existing
Settings dropdown:

- /tenants/{id}/features — original page restored (handler/routes/template)
- /tenants/{id}/defaults — new page hosting only the locale override

Drops the gear icon between tenant chip and user from the top bar; the
Settings dropdown is the only entry point. Also removes the Settings
quick action from the tenant home dashboard.

The locale infrastructure (Tenant.defaultLocale, SetTenantDefaultLocale
command, TenantLocaleResolver, LocaleProperties, migration
V20260519142326__core_tenant_default_locale.sql) is unchanged and now
consumed by the new DefaultsHandler — which follows the same simple
form-POST + 303 pattern as FeatureHandler instead of the HTMX-fragment
approach the deleted settings page used.

Refs #353, #289
