+++ apps/epistola/src/main/kotlin/app/epistola/suite/handlers/FeatureRoutes.kt
@@ -1,0 +1,18 @@
+package app.epistola.suite.handlers
+
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.function.RouterFunction
+import org.springframework.web.servlet.function.ServerResponse
+import org.springframework.web.servlet.function.router
+
+@Configuration
+class FeatureRoutes(private val handler: FeatureHandler) {
+    @Bean
+    fun featureRouterFunction(): RouterFunction<ServerResponse> = router {
+        "/tenants/{tenantId}/features".nest {
+            GET("", handler::featureToggles)
+            POST("", handler::saveFeatureToggles)
+        }
+    }
+}
