+++ apps/epistola/src/main/resources/templates/layout/nav.html
@@ -122,6 +122,29 @@
                         </div>
                     </div>
 
+                    <!-- Settings dropdown (managers only) -->
+                    <div th:if="${auth.has('TENANT_SETTINGS')}"
+                         class="app-nav-dropdown"
+                         th:classappend="${#lists.contains({'features','defaults'}, activeNavSection)} ? ' active' : ''">
+                        <button class="app-nav-dropdown-trigger" type="button"
+                                aria-expanded="false" aria-haspopup="true">
+                            Settings
+                            <th:block th:replace="~{fragments/icon :: icon(name='chevron-down', class='ep-icon ep-icon-sm')}" />
+                        </button>
+                        <div class="app-nav-dropdown-menu" role="menu">
+                            <a th:href="@{/tenants/{id}/features(id=${tenantId})}"
+                               class="app-nav-dropdown-item" role="menuitem"
+                               th:classappend="${activeNavSection == 'features'} ? ' active' : ''">
+                                Features
+                            </a>
+                            <a th:href="@{/tenants/{id}/defaults(id=${tenantId})}"
+                               class="app-nav-dropdown-item" role="menuitem"
+                               th:classappend="${activeNavSection == 'defaults'} ? ' active' : ''">
+                                Defaults
+                            </a>
+                        </div>
+                    </div>
+
                 </div>
 
                 <!-- Mobile hamburger toggle -->
@@ -134,15 +157,6 @@
                 <a th:href="@{/}" class="app-nav-tenant" title="Switch tenant">
                     <th:block th:replace="~{fragments/icon :: icon(name='building-2', class='ep-icon ep-icon-sm')}" />
                     <span th:text="${tenantName} ?: 'Tenant'">Tenant</span>
-                </a>
-
-                <a th:if="${tenantId != null and auth?.has('TENANT_SETTINGS')}"
-                   th:href="@{/tenants/{id}/settings(id=${tenantId})}"
-                   class="app-nav-icon-link"
-                   th:classappend="${activeNavSection == 'settings'} ? ' active' : ''"
-                   title="Settings"
-                   aria-label="Settings">
-                    <th:block th:replace="~{fragments/icon :: icon(name='settings', class='ep-icon ep-icon-sm')}" />
                 </a>
 
                 <div class="app-nav-user" sec:authorize="isAuthenticated()">
@@ -205,8 +219,10 @@
                 </div>
                 <div th:if="${auth.has('TENANT_SETTINGS')}" class="app-nav-mobile-group">
                     <div class="app-nav-mobile-label">Settings</div>
-                    <a th:href="@{/tenants/{id}/settings(id=${tenantId})}" class="app-nav-mobile-item"
-                       th:classappend="${activeNavSection == 'settings'} ? ' active' : ''" role="menuitem">Settings</a>
+                    <a th:href="@{/tenants/{id}/features(id=${tenantId})}" class="app-nav-mobile-item"
+                       th:classappend="${activeNavSection == 'features'} ? ' active' : ''" role="menuitem">Features</a>
+                    <a th:href="@{/tenants/{id}/defaults(id=${tenantId})}" class="app-nav-mobile-item"
+                       th:classappend="${activeNavSection == 'defaults'} ? ' active' : ''" role="menuitem">Defaults</a>
                 </div>
                 <div class="app-nav-mobile-footer">
                     <a th:href="@{/}" class="app-nav-mobile-item" role="menuitem">
