+++ apps/epistola/src/main/resources/templates/tenants/home.html
@@ -87,10 +87,6 @@
                         <th:block th:replace="~{fragments/icon :: icon(name='zap', class='ep-icon')}" />
                         Generate Document
                     </a>
-                    <a th:href="@{/tenants/{id}/settings(id=${tenant.id})}" class="ep-btn ep-btn-outline">
-                        <th:block th:replace="~{fragments/icon :: icon(name='settings', class='ep-icon')}" />
-                        Settings
-                    </a>
                 </div>
             </div>
         </div>
