--- apps/epistola/src/main/resources/templates/tenants/settings.html
@@ -1,113 +1,0 @@
-<!DOCTYPE html>
-<html lang="en" xmlns:th="http://www.thymeleaf.org">
-<body>
-    <th:block th:fragment="content">
-        <nav class="breadcrumb">
-            <a th:href="@{/tenants/{tenantId}(tenantId=${tenantId})}" th:text="${tenant.name}">Tenant</a>
-            &raquo; <span>Settings</span>
-        </nav>
-
-        <div class="page-header">
-            <div>
-                <h1>Settings</h1>
-                <p class="subtitle" th:text="'Tenant: ' + ${tenant.id}">Tenant: tenant-id</p>
-            </div>
-        </div>
-
-        <section class="card">
-            <div class="card-header">
-                <h2>Default locale</h2>
-                <p class="subtitle">
-                    Used to format dates and numbers in generated documents.
-                    Leave blank to use the application default.
-                </p>
-            </div>
-            <div class="card-content">
-                <div th:replace="~{tenants/settings :: settings-form}"></div>
-            </div>
-        </section>
-
-        <section class="card">
-            <div class="card-header">
-                <h2>Feature toggles</h2>
-                <p class="subtitle">
-                    Per-tenant overrides of the application-wide feature defaults.
-                </p>
-            </div>
-            <div class="card-content">
-                <div th:replace="~{tenants/settings :: features-form}"></div>
-            </div>
-        </section>
-    </th:block>
-
-    <th:block th:fragment="settings-form">
-        <form id="settings-form"
-              th:action="@{/tenants/{tenantId}/settings/locale(tenantId=${tenantId})}"
-              method="post"
-              th:hx-post="@{/tenants/{tenantId}/settings/locale(tenantId=${tenantId})}"
-              hx-target="#settings-form"
-              hx-swap="outerHTML">
-
-            <p class="form-hint">
-                <strong>Effective locale:</strong>
-                <span th:text="${effectiveLocale}">en-US</span>
-                <span th:if="${tenant.defaultLocale != null}"
-                      th:text="'(tenant override)'">(tenant override)</span>
-                <span th:if="${tenant.defaultLocale == null}"
-                      th:text="'(application default: ' + ${applicationDefault} + ')'">
-                    (application default: en-US)
-                </span>
-            </p>
-
-            <div class="form-group" th:classappend="${errors?.containsKey('locale')} ? 'error' : ''">
-                <label class="ep-label" for="locale">Locale override</label>
-                <select id="locale" name="locale" class="ep-input">
-                    <option value="" th:selected="${tenant.defaultLocale == null}">
-                        — Use application default —
-                    </option>
-                    <option th:each="entry : ${entries}"
-                            th:value="${entry.code}"
-                            th:text="${entry.label + ' (' + entry.code + ')'}"
-                            th:selected="${entry.code == tenant.defaultLocale}">en-US</option>
-                </select>
-                <span class="form-error"
-                      th:if="${errors?.containsKey('locale')}"
-                      th:text="${errors.locale}"></span>
-            </div>
-
-            <div class="form-actions">
-                <button type="submit" class="ep-btn ep-btn-primary">Save</button>
-            </div>
-        </form>
-    </th:block>
-
-    <th:block th:fragment="features-form">
-        <form id="features-form"
-              th:action="@{/tenants/{tenantId}/settings/features(tenantId=${tenantId})}"
-              method="post"
-              th:hx-post="@{/tenants/{tenantId}/settings/features(tenantId=${tenantId})}"
-              hx-target="#features-form"
-              hx-swap="outerHTML">
-
-            <div class="feature-toggle-item" th:each="feature : ${features}">
-                <div class="feature-toggle-info">
-                    <div class="feature-toggle-name" th:text="${feature['key']}">feature-key</div>
-                    <div class="feature-toggle-description"
-                         th:if="${feature['description'] != ''}"
-                         th:text="${feature['description']}"></div>
-                </div>
-                <label class="toggle-switch">
-                    <input type="checkbox"
-                           th:name="${feature['key']}"
-                           th:checked="${feature['enabled']}">
-                    <span class="toggle-switch-slider"></span>
-                </label>
-            </div>
-
-            <div class="form-actions">
-                <button type="submit" class="ep-btn ep-btn-primary">Save feature toggles</button>
-            </div>
-        </form>
-    </th:block>
-</body>
-</html>
